moved micropython.mk to the top level directory
this allow the module to be used in combination with other modules without
modifying it at all.

removed -DMODULE_ST7789_ENABLED=1 since that does not always apply, this should
be defined in the configuration of the project that uses this module.

The module directory may contain modules that only apply to specific ports,
forcing the module to be enabled in micropython.mk, defeats the purpose of
having a define to enable it.